LuisaViaRoma-InventioHub/backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
boto3==1.38.18
botocore==1.38.18
cached-property==2.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==45.0.2
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-simple-history==3.8.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
oauthlib==3.2.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.31.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
pyparsing==3.2.3
python-datauri==3.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.12.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.1.1
urllib3==2.4.0
uWSGI
